Favicon and kitchen stuff

+++ prototype/pages/kitchen.vue
@@ -1,21 +1,22 @@
 <template>
   <div class="kitchen-room">
     <nav class="kitchen-nav">
-      <button @click="handleAdd">Add a studio</button>
+      <button @click="isAddingSupportedStudio = !isAddingSupportedStudio">
+        {{ isAddingSupportedStudio ? "I'm done" : "Add a studio" }}
+      </button>
     </nav>
 
     <div v-if="isAddingSupportedStudio" class="kitchen-add">
       <div class="kitchen-add-header">
         <h2>Add a studio</h2>
-        <button @click="isAddingSupportedStudio = false">Done</button>
       </div>
       <div>
         <label for="query">Search for a place with its name and/or address</label>
       </div>
-      <div class="kitchen-add-form">
+      <form class="kitchen-add-form" onsubmit="return false">
         <input type="text" v-model="mapsSearchQuery" placeholder="" name="query"/>
-        <button @click="handleSearchMaps">Search Google Maps</button>
-      </div>
+        <button type="submit" @click="handleSearchMaps">Search Google Maps</button>
+      </form>
 
       <div v-if="mapsSearchCandidates.length">
         <h3>Results</h3>
@@ -23,7 +24,8 @@
           <li v-for="candidate in mapsSearchCandidates" :key="candidate.place_id">
             <p>{{ candidate.name }}</p>
             <p>{{ candidate.formatted_address }}</p>
-            <button @click="handleAddStudio(candidate.place_id)">Add this studio</button>
+            <button v-if="isAdded(candidate.place_id)" disabled>Added</button>
+            <button v-else @click="handleAddStudio(candidate.place_id)">Add this studio</button>
           </li>
         </ol>
       </div>
@@ -31,9 +33,11 @@
 
     <div class="kitchen-table">
       <div class="kitchen-table-header">
-        <p>The database will update from Google Maps data every day.</p>
+        <p>Hi. Welcome to the Supported Studios kitchen. Here, you can add or delete new places, and links for them.</p>
+        <p>Just so you know, if you add places to the Google Maps list, it won't show up here. But if you add a place here, the kitchen
+          will check Google Maps every day to make sure the name, location, contact info, and status is up to date.</p>
       </div>
-      <PlaceTable :data="data" :is-kitchen-mode="true" @deleteStudio="handleDeleteStudio"/>
+      <PlaceTable :data="places" :is-kitchen-mode="true" @deleteStudio="handleDeleteStudio" @update-place="handleUpdatePlace"/>
     </div>
   </div>
 </template>
@@ -41,9 +45,6 @@
 <script setup>
 import { ref } from 'vue';
 const isAddingSupportedStudio = ref(false)
-const handleAdd = () => {
-  isAddingSupportedStudio.value = true
-}
 const mapsSearchQuery = ref('')
 const mapsSearchCandidates = ref([])
 const handleSearchMaps = async () => {
@@ -62,10 +63,14 @@ const handleAddStudio = async (placeId) => {
     body: { placeId }
   })
   console.log(data)
+  places.value.push(data.value)
 }
 
 const handleDeleteStudio = async(placeId) => {
   console.log(placeId)
+  if (!confirm("Just making sure you didn't click this by mistake. Are you sure you want to delete this studio?")) {
+    return
+  }
   const { data } = await useFetch('/api/place', {
     method: 'DELETE',
     body: { placeId }
@@ -73,5 +78,18 @@ const handleDeleteStudio = async(placeId) => {
   console.log(data)
 }
 
+const handleUpdatePlace = async(placeData) => {
+  console.log(placeData)
+  console.log(places.value)
+  placeData.links = JSON.parse(placeData.links || '[]')
+  const placeIndex = places.value.findIndex(place => place.place_id === placeData.place_id)
+  places.value.splice(placeIndex, 1, placeData)
+}
+
+const isAdded = (placeId) => {
+  return places.value.findIndex(place => place.place_id === placeId) > -1
+}
+
 const { data } = await useFetch('/api/place/all')
+const places = ref(data.value)
 </script>
